// types/swapi.d.ts
export interface Film {
  /**
   * The title of this film.
   */
  title: string;
  /**
   * The episode number of this film.
   */
  episode_id: number;
  /**
   * The opening crawl text at the beginning of this film.
   */
  opening_crawl: string;
  /**
   * The director of this film.
   */
  director: string;
  /**
   * The producer(s) of this film.
   */
  producer: string;
  /**
   * The release date at original creator country.
   */
  release_date: string;
  /**
   * The people resources featured within this film.
   */
  characters: string[];
  /**
   * The planet resources featured within this film.
   */
  planets: string[];
  /**
   * The starship resources featured within this film.
   */
  starships: string[];
  /**
   * The vehicle resources featured within this film.
   */
  vehicles: string[];
  /**
   * The species resources featured within this film.
   */
  species: string[];
  /**
   * The url of this resource
   */
  url: string;
  /**
   * The ISO 8601 date format of the time that this resource was created.
   */
  created: string;
  /**
   * the ISO 8601 date format of the time that this resource was edited.
   */
  edited: string;
  [k: string]: unknown;
}

/**
 * A person within the Star Wars universe
 */
export interface People {
  /**
   * The name of this person.
   */
  name: string;
  /**
   * The height of this person in meters.
   */
  height: string;
  /**
   * The mass of this person in kilograms.
   */
  mass: string;
  /**
   * The hair color of this person.
   */
  hair_color: string;
  /**
   * The skin color of this person.
   */
  skin_color: string;
  /**
   * The eye color of this person.
   */
  eye_color: string;
  /**
   * The birth year of this person. BBY (Before the Battle of Yavin) or ABY (After the Battle of Yavin).
   */
  birth_year: string;
  /**
   * The gender of this person (if known).
   */
  gender: string;
  /**
   * The url of the planet resource that this person was born on.
   */
  homeworld: string;
  /**
   * An array of urls of film resources that this person has been in.
   */
  films: string[];
  /**
   * The url of the species resource that this person is.
   */
  species: string[];
  /**
   * An array of vehicle resources that this person has piloted
   */
  vehicles: string[];
  /**
   * An array of starship resources that this person has piloted
   */
  starships: string[];
  /**
   * The url of this resource
   */
  url: string;
  /**
   * The ISO 8601 date format of the time that this resource was created.
   */
  created: string;
  /**
   * the ISO 8601 date format of the time that this resource was edited.
   */
  edited: string;
  [k: string]: unknown;
}

/**
 * A planet.
 */
export interface Planet {
  /**
   * The name of this planet.
   */
  name: string;
  /**
   * The diameter of this planet in kilometers.
   */
  diameter: string;
  /**
   * The number of standard hours it takes for this planet to complete a single rotation on its axis.
   */
  rotation_period: string;
  /**
   * The number of standard days it takes for this planet to complete a single orbit of its local star.
   */
  orbital_period: string;
  /**
   * A number denoting the gravity of this planet. Where 1 is normal.
   */
  gravity: string;
  /**
   * The average populationof sentient beings inhabiting this planet.
   */
  population: string;
  /**
   * The climate of this planet. Comma-seperated if diverse.
   */
  climate: string;
  /**
   * the terrain of this planet. Comma-seperated if diverse.
   */
  terrain: string;
  /**
   * The percentage of the planet surface that is naturally occuring water or bodies of water.
   */
  surface_water: string;
  /**
   * An array of Film URL Resources that this planet has appeared in.
   */
  films: string[];
  /**
   * An array of People URL Resources that live on this planet.
   */
  residents: string[];
  /**
   * The hypermedia URL of this resource.
   */
  url: string;
  /**
   * The ISO 8601 date format of the time that this resource was created.
   */
  created: string;
  /**
   * the ISO 8601 date format of the time that this resource was edited.
   */
  edited: string;
  [k: string]: unknown;
}